markers/requirements.txt
Django==1.9.6
django-configurations==1.0
factory-boy==2.6.0
fake-factory==0.5.3
py==1.4.31
pytest==2.8.2
pytest-django==2.9.1
whitenoise==2.0.6
